handlebars working as expected and modal working on att
