+++ models/student.js
@@ -1,26 +1,33 @@
 module.exports = (sequelize, DataTypes) => {
-  const Student = sequelize.define('Student', {
-    student_name: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      validate: {
-        len: [1],
+  const Student = sequelize.define(
+    'Student',
+    {
+      student_name: {
+        type: DataTypes.STRING,
+        allowNull: false,
+        validate: {
+          len: [1],
+        },
+      },
+      teacher_main: {
+        type: DataTypes.STRING,
+        allowNull: false,
+        validate: {
+          len: [1],
+        },
+      },
+      teacher_email: {
+        type: DataTypes.STRING,
+        allowNull: false,
+        validate: {
+          len: [1],
+        },
       },
     },
-    teacher_main: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      validate: {
-        len: [1],
-      },
-    },
-    teacher_email: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      validate: {
-        len: [1],
-      },
-    },
-  });
+    {
+      freezeTableName: true,
+      timestamps: false,
+    }
+  );
   return Student;
 };
