+++ models/user.js
@@ -2,21 +2,28 @@
 const bcrypt = require('bcryptjs');
 
 module.exports = (sequelize, DataTypes) => {
-  const User = sequelize.define('User', {
-    email: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true,
-      validate: {
-        isEmail: true,
+  const User = sequelize.define(
+    'User',
+    {
+      email: {
+        type: DataTypes.STRING,
+        allowNull: false,
+        unique: true,
+        validate: {
+          isEmail: true,
+        },
+      },
+
+      password: {
+        type: DataTypes.STRING,
+        allowNull: false,
       },
     },
-
-    password: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-  });
+    {
+      freezeTableName: true,
+      timestamps: false,
+    }
+  );
 
   User.prototype.validPassword = function (password) {
     return bcrypt.compareSync(password, this.password);
